Move PostJSONWithTimeout to platform/http package
