+++ cmd/fleet/serve.go
@@ -1813,9 +1813,6 @@ func createActivityBoundedContext(svc fleet.Service, dbConns *common_mysql.DBCon
 		dbConns,
 		activityAuthorizer,
 		activityACLAdapter,
-		func(ctx context.Context, url string, payload any) error {
-			return server.PostJSONWithTimeout(ctx, url, payload, logger)
-		},
 		logger,
 	)
 	// Create auth middleware for activity bounded context
