+++ server/activity/bootstrap/bootstrap.go
@@ -20,11 +20,10 @@ func New(
 	dbConns *platform_mysql.DBConnections,
 	authorizer platform_authz.Authorizer,
 	providers activity.DataProviders,
-	webhookSendFn activity.WebhookSendFunc,
 	logger *slog.Logger,
 ) (api.Service, func(authMiddleware endpoint.Middleware) eu.HandlerRoutesFunc) {
 	ds := mysql.NewDatastore(dbConns, logger)
-	svc := service.NewService(authorizer, ds, providers, webhookSendFn, logger)
+	svc := service.NewService(authorizer, ds, providers, logger)
 
 	routesFn := func(authMiddleware endpoint.Middleware) eu.HandlerRoutesFunc {
 		return service.GetRoutes(svc, authMiddleware)
