+++ server/activity/internal/tests/suite_test.go
@@ -1,7 +1,6 @@
 package tests
 
 import (
-	"context"
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
@@ -38,8 +37,7 @@ func setupIntegrationTest(t *testing.T) *integrationTestSuite {
 	providers := newMockDataProviders()
 
 	// Create service
-	noopWebhookSend := func(_ context.Context, _ string, _ any) error { return nil }
-	svc := service.NewService(authorizer, ds, providers, noopWebhookSend, tdb.Logger)
+	svc := service.NewService(authorizer, ds, providers, tdb.Logger)
 
 	// Create router with routes
 	router := mux.NewRouter()
