+++ server/fleet/errors.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"time"
 
+	platform_errors "github.com/fleetdm/fleet/v4/server/platform/errors"
 	platform_http "github.com/fleetdm/fleet/v4/server/platform/http"
 	"github.com/rs/zerolog"
 )
@@ -51,8 +52,8 @@ type ErrWithLogFields = platform_http.ErrWithLogFields
 // ErrWithRetryAfter is an alias for platform_http.ErrWithRetryAfter.
 type ErrWithRetryAfter = platform_http.ErrWithRetryAfter
 
-// ErrWithIsClientError is an alias for platform_http.ErrWithIsClientError.
-type ErrWithIsClientError = platform_http.ErrWithIsClientError
+// ErrWithIsClientError is an alias for platform_errors.ErrWithIsClientError.
+type ErrWithIsClientError = platform_errors.ErrWithIsClientError
 
 type invalidArgWithStatusError struct {
 	InvalidArgumentError
@@ -400,7 +401,7 @@ func GetJSONUnknownField(err error) *string {
 }
 
 // Cause returns the root error in err's chain.
-var Cause = platform_http.Cause
+var Cause = platform_errors.Cause
 
 // FleetdError is an error that can be reported by any of the fleetd
 // components.
