+++ server/service/activities_test.go
@@ -2,61 +2,15 @@ package service
 
 import (
 	"context"
-	"encoding/json"
-	"log/slog"
-	"net/http"
-	"net/http/httptest"
 	"testing"
-	"time"
 
-	fleetserver "github.com/fleetdm/fleet/v4/server"
-	"github.com/fleetdm/fleet/v4/server/activity"
 	activity_api "github.com/fleetdm/fleet/v4/server/activity/api"
-	activity_bootstrap "github.com/fleetdm/fleet/v4/server/activity/bootstrap"
 	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mock"
 	"github.com/fleetdm/fleet/v4/server/ptr"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-// webhookTestProviders implements activity.DataProviders for webhook tests.
-type webhookTestProviders struct {
-	getWebhookConfig func() (*activity.ActivitiesWebhookSettings, error)
-}
-
-func (p *webhookTestProviders) GetActivitiesWebhookConfig(_ context.Context) (*activity.ActivitiesWebhookSettings, error) {
-	return p.getWebhookConfig()
-}
-
-func (p *webhookTestProviders) ActivateNextUpcomingActivity(_ context.Context, _ uint, _ string) error {
-	return nil
-}
-
-func (p *webhookTestProviders) MaskSecretURLParams(rawURL string) string { return rawURL }
-func (p *webhookTestProviders) MaskURLError(err error) error             { return err }
-func (p *webhookTestProviders) UsersByIDs(_ context.Context, _ []uint) ([]*activity.User, error) {
-	return nil, nil
-}
-func (p *webhookTestProviders) FindUserIDs(_ context.Context, _ string) ([]uint, error) {
-	return nil, nil
-}
-func (p *webhookTestProviders) GetHostLite(_ context.Context, _ uint) (*activity.Host, error) {
-	return nil, nil
-}
-
-type ActivityTypeTest struct {
-	Name string `json:"name"`
-}
-
-func (a ActivityTypeTest) ActivityName() string {
-	return "test_activity"
-}
-
-func (a ActivityTypeTest) Documentation() (activity string, details string, detailsExample string) {
-	return "test_activity", "test_activity", "test_activity"
-}
 
 func Test_logRoleChangeActivities(t *testing.T) {
 	tests := []struct {
@@ -147,220 +101,6 @@ func Test_logRoleChangeActivities(t *testing.T) {
 			require.Equal(t, tt.expectActivities, activities)
 		})
 	}
-}
-
-func TestActivityWebhooks(t *testing.T) {
-	ds := new(mock.Store)
-	opts := &TestServerOpts{}
-	svc, ctx := newTestService(t, ds, nil, nil, opts)
-	var webhookBody = fleet.ActivityWebhookPayload{}
-	webhookChannel := make(chan struct{}, 1)
-	fail429 := false
-	startMockServer := func(t *testing.T) string {
-		// create a test http server
-		srv := httptest.NewServer(
-			http.HandlerFunc(
-				func(w http.ResponseWriter, r *http.Request) {
-					webhookBody = fleet.ActivityWebhookPayload{}
-					if r.Method != "POST" {
-						w.WriteHeader(http.StatusMethodNotAllowed)
-						return // don't send the channel signal
-					}
-					switch r.URL.Path {
-					case "/ok":
-						err := json.NewDecoder(r.Body).Decode(&webhookBody)
-						if err != nil {
-							t.Log(err)
-							w.WriteHeader(http.StatusBadRequest)
-						}
-					case "/error":
-						webhookBody.Type = "error" // to check for testing
-						w.WriteHeader(http.StatusTeapot)
-					case "/429":
-						// Only the first request will fail
-						fail429 = !fail429
-						if fail429 {
-							w.WriteHeader(http.StatusTooManyRequests)
-							return // don't send the channel signal
-						}
-						err := json.NewDecoder(r.Body).Decode(&webhookBody)
-						if err != nil {
-							t.Log(err)
-							w.WriteHeader(http.StatusBadRequest)
-						}
-					default:
-						w.WriteHeader(http.StatusNotFound)
-						return // don't send the channel signal
-					}
-					webhookChannel <- struct{}{}
-				},
-			),
-		)
-		t.Cleanup(srv.Close)
-		return srv.URL
-	}
-	mockUrl := startMockServer(t)
-	testUrl := mockUrl
-
-	// Wire a real activity bounded context service as delegate so that webhook
-	// firing (which lives in the bounded context) is exercised. The mock still
-	// captures invocations and the user for assertions.
-	providers := &webhookTestProviders{
-		getWebhookConfig: func() (*activity.ActivitiesWebhookSettings, error) {
-			return &activity.ActivitiesWebhookSettings{
-				Enable:         true,
-				DestinationURL: testUrl,
-			}, nil
-		},
-	}
-	discardLogger := slog.New(slog.DiscardHandler)
-	realActivitySvc := activity_bootstrap.NewForUnitTests(providers, func(ctx context.Context, url string, payload any) error {
-		return fleetserver.PostJSONWithTimeout(ctx, url, payload, discardLogger)
-	}, discardLogger)
-	opts.ActivityMock.Delegate = realActivitySvc
-
-	var activityUser *activity_api.User
-	opts.ActivityMock.NewActivityFunc = func(_ context.Context, user *activity_api.User, _ activity_api.ActivityDetails) error {
-		activityUser = user
-		return nil
-	}
-
-	tests := []struct {
-		name    string
-		user    *fleet.User
-		url     string
-		doError bool
-	}{
-		{
-			name: "nil user",
-			url:  mockUrl + "/ok",
-			user: nil,
-		},
-		{
-			name: "real user",
-			url:  mockUrl + "/ok",
-			user: &fleet.User{
-				ID:    1,
-				Name:  "testUser",
-				Email: "testUser@example.com",
-			},
-		},
-		{
-			name:    "error",
-			url:     mockUrl + "/error",
-			doError: true,
-		},
-		{
-			name: "429",
-			url:  mockUrl + "/429",
-			user: &fleet.User{
-				ID:    2,
-				Name:  "testUser2",
-				Email: "testUser2@example.com",
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(
-			tt.name, func(t *testing.T) {
-				opts.ActivityMock.NewActivityFuncInvoked = false
-				testUrl = tt.url
-				startTime := time.Now()
-				act := ActivityTypeTest{Name: tt.name}
-				err := svc.NewActivity(ctx, tt.user, act)
-				require.NoError(t, err)
-				select {
-				case <-time.After(1 * time.Second):
-					t.Error("timeout")
-				case <-webhookChannel:
-					if tt.doError {
-						assert.Equal(t, "error", webhookBody.Type)
-					} else {
-						endTime := time.Now()
-						assert.False(
-							t, webhookBody.Timestamp.Before(startTime), "timestamp %s is before start time %s",
-							webhookBody.Timestamp.String(), startTime.String(),
-						)
-						assert.False(t, webhookBody.Timestamp.After(endTime))
-						if tt.user == nil {
-							assert.Nil(t, webhookBody.ActorFullName)
-							assert.Nil(t, webhookBody.ActorID)
-							assert.Nil(t, webhookBody.ActorEmail)
-						} else {
-							require.NotNil(t, webhookBody.ActorFullName)
-							assert.Equal(t, tt.user.Name, *webhookBody.ActorFullName)
-							require.NotNil(t, webhookBody.ActorID)
-							assert.Equal(t, tt.user.ID, *webhookBody.ActorID)
-							require.NotNil(t, webhookBody.ActorEmail)
-							assert.Equal(t, tt.user.Email, *webhookBody.ActorEmail)
-						}
-						assert.Equal(t, act.ActivityName(), webhookBody.Type)
-						var details map[string]string
-						require.NoError(t, json.Unmarshal(*webhookBody.Details, &details))
-						assert.Len(t, details, 1)
-						assert.Equal(t, tt.name, details["name"])
-					}
-				}
-				require.True(t, opts.ActivityMock.NewActivityFuncInvoked)
-				if tt.user == nil {
-					assert.Nil(t, activityUser)
-				} else {
-					require.NotNil(t, activityUser)
-					assert.Equal(t, tt.user.ID, activityUser.ID)
-					assert.Equal(t, tt.user.Name, activityUser.Name)
-					assert.Equal(t, tt.user.Email, activityUser.Email)
-				}
-			},
-		)
-	}
-}
-
-func TestActivityWebhooksDisabled(t *testing.T) {
-	ds := new(mock.Store)
-	opts := &TestServerOpts{}
-	svc, ctx := newTestService(t, ds, nil, nil, opts)
-	startMockServer := func(t *testing.T) string {
-		// create a test http server
-		srv := httptest.NewServer(
-			http.HandlerFunc(
-				func(w http.ResponseWriter, r *http.Request) {
-					t.Error("should not be called")
-				},
-			),
-		)
-		t.Cleanup(srv.Close)
-		return srv.URL
-	}
-	mockUrl := startMockServer(t)
-
-	ds.AppConfigFunc = func(ctx context.Context) (*fleet.AppConfig, error) {
-		return &fleet.AppConfig{
-			WebhookSettings: fleet.WebhookSettings{
-				ActivitiesWebhook: fleet.ActivitiesWebhookSettings{
-					Enable:         false,
-					DestinationURL: mockUrl,
-				},
-			},
-		}, nil
-	}
-	var activityUser *activity_api.User
-	opts.ActivityMock.NewActivityFunc = func(_ context.Context, user *activity_api.User, _ activity_api.ActivityDetails) error {
-		activityUser = user
-		return nil
-	}
-	activity := ActivityTypeTest{Name: "no webhook"}
-	user := &fleet.User{
-		ID:    1,
-		Name:  "testUser",
-		Email: "testUser@example.com",
-	}
-	require.NoError(t, svc.NewActivity(ctx, user, activity))
-	require.True(t, opts.ActivityMock.NewActivityFuncInvoked)
-	require.NotNil(t, activityUser)
-	assert.Equal(t, user.ID, activityUser.ID)
-	assert.Equal(t, user.Name, activityUser.Name)
-	assert.Equal(t, user.Email, activityUser.Email)
 }
 
 func TestCancelHostUpcomingActivityAuth(t *testing.T) {
