+++ server/utils.go
@@ -1,24 +1,16 @@
 package server
 
 import (
-	"bytes"
-	"context"
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/base64"
-	"encoding/json"
 	"encoding/pem"
 	"errors"
 	"fmt"
 	"html/template"
-	"io"
-	"log/slog"
-	"net/http"
 	"strings"
-	"time"
 
-	"github.com/fleetdm/fleet/v4/pkg/fleethttp"
 	"github.com/fleetdm/fleet/v4/server/bindata"
 	platformhttp "github.com/fleetdm/fleet/v4/server/platform/http"
 )
@@ -45,62 +37,10 @@ func GenerateRandomURLSafeText(keySize int) (string, error) {
 	return base64.URLEncoding.EncodeToString(key), nil
 }
 
-func httpSuccessStatus(statusCode int) bool {
-	return statusCode >= 200 && statusCode <= 299
-}
-
-// errWithStatus is an error with a particular status code.
-type errWithStatus struct {
-	err        string
-	statusCode int
-}
-
-// Error implements the error interface
-func (e *errWithStatus) Error() string {
-	return e.err
-}
-
-// StatusCode implements the StatusCoder interface for returning custom status codes.
-func (e *errWithStatus) StatusCode() int {
-	return e.statusCode
-}
-
-func PostJSONWithTimeout(ctx context.Context, url string, v any, logger *slog.Logger) error {
-	jsonBytes, err := json.Marshal(v)
-	if err != nil {
-		return err
-	}
-
-	client := fleethttp.NewClient(fleethttp.WithTimeout(30 * time.Second))
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, url, bytes.NewBuffer(jsonBytes))
-	if err != nil {
-		return err
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to POST to %s: %s, request-size=%d", MaskSecretURLParams(url), MaskURLError(err), len(jsonBytes))
-	}
-	defer resp.Body.Close()
-
-	if !httpSuccessStatus(resp.StatusCode) {
-		body, _ := io.ReadAll(resp.Body)
-		bodyStr := string(body)
-		if len(bodyStr) > 512 {
-			bodyStr = bodyStr[:512]
-		}
-		logger.DebugContext(ctx, "non-success response from POST",
-			"url", MaskSecretURLParams(url),
-			"status_code", resp.StatusCode,
-			"body", bodyStr,
-		)
-		return &errWithStatus{err: fmt.Sprintf("error posting to %s", MaskSecretURLParams(url)), statusCode: resp.StatusCode}
-	}
-
-	return nil
-}
+// PostJSONWithTimeout marshals v as JSON and POSTs it to the given URL with a 30-second timeout.
+//
+// Deprecated: Use github.com/fleetdm/fleet/v4/server/platform/http.PostJSONWithTimeout instead.
+var PostJSONWithTimeout = platformhttp.PostJSONWithTimeout
 
 // MaskSecretURLParams masks URL query values if the query param name includes "secret", "token",
 // "key", "password". It accepts a raw string and returns a redacted string if the raw string is
